+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Output summary uses `\t` whitespaces for easier copying.
   [#21](https://github.com/netsec-ethz/scion-java-multiping/pull/21)
+- Output uses more `\t` + drop ICMP out if disabled + print runtime
+  [#23](https://github.com/netsec-ethz/scion-java-multiping/pull/23)
 
 ### Fixed