+++ CHANGELOG.md
@@ -19,6 +19,7 @@ TODO
   - Use \t in summary output
 - Investigate spurious 0-0:0 ASes in PingAll output
 
+## [0.6.0] - 2025-12-12
 
 ### Added
 
@@ -126,7 +127,8 @@ TODO
 
 - Nothing
 
-[Unreleased]: https://github.com/netsec-ethz/scion-java-multiping/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/netsec-ethz/scion-java-multiping/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/netsec-ethz/scion-java-multiping/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/netsec-ethz/scion-java-multiping/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/netsec-ethz/scion-java-multiping/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/netsec-ethz/scion-java-multiping/compare/v0.2.0...v0.3.0